+++ CHANGES.md
@@ -1,6 +1,8 @@
 HEAD
 ----
 
+* Added `:all_accept_arguments` to Slop configuration options, this saves
+  having to specify that every option takes an argument
 * Added `Slop#to_struct` for building new classes from options
 
 2.0.0 (2011-07-07)